Fix: Local storage bug on mobile
